+++ tools/base/BUILD
@@ -5,8 +5,13 @@ licenses(["notice"])  # Apache 2
 
 envoy_package()
 
-envoy_py_library("tools.base.checker")
-
 envoy_py_library("tools.base.runner")
 
 envoy_py_library("tools.base.utils")
+
+envoy_py_library(
+    "tools.base.checker",
+    deps = [
+        ":runner",
+    ],
+)
